Added validations in services

+++ src/main/java/com/hoodiesbackend/services/cart/cartItem/CartItemService.java
@@ -1,6 +1,7 @@
 package com.hoodiesbackend.services.cart.cartItem;
 
 import com.hoodiesbackend.entities.cart.CartItem.CartItem;
+import com.hoodiesbackend.exceptions.BadRequestException;
 import com.hoodiesbackend.exceptions.NotFoundException;
 import com.hoodiesbackend.repositories.cart.CartItemRepository;
 import org.springframework.stereotype.Service;
@@ -25,24 +26,33 @@ public class CartItemService {
     }
 
     public List<CartItem> findAllByCartId(Long cartId){
+        if(cartId < 1) throw new BadRequestException("Cart Id should be > 0");
         return cartItemRepository.findAllByCartId(cartId);
     }
 
     public CartItem findItemForCart(Long cartId, Long itemId){
+        if(cartId < 1 || itemId < 1) throw new BadRequestException("Cart Id and Item id should be > 0");
+
         return cartItemRepository.findCartItemByCartIdAndId(cartId, itemId)
                 .orElseThrow(() -> new NotFoundException("Item not found in cart!"));
     }
 
     public CartItem findItem(Long cartId, Long productId){
+        if(cartId < 1 || productId < 1) throw new BadRequestException("Cart Id and Product id should be > 0");
+
         return cartItemRepository.findCartItemByCartIdAndProductId(cartId, productId)
                 .orElseThrow(() -> new NotFoundException("Item not found in cart!"));
     }
 
     public boolean delete(Long cartId, Long itemId){
+        if(cartId < 1 || itemId < 1) throw new BadRequestException("Cart Id and Item id should be > 0");
+
         return cartItemRepository.deleteCartItemByCartIdAndId(cartId, itemId) > 0;
     }
 
     public Boolean contains(Long cartId, Long productId){
+        if(cartId < 1 || productId < 1) throw new BadRequestException("Cart Id and Product id should be > 0");
+
         return cartItemRepository.existsCartItemByCartIdAndProductId(cartId, productId);
     }
 }
